Add AUTHORS, CONTRIBUTING, LICENSE, RATIONALE, README, and SUMMARY
This commit represent a first stab at a template for open source
software projects under the Zambezi banner. It adds sensible
defaults for most things, and elsewhere specifies where things
should change.

The rationale can probably further explain *why* some choices were
made (such as why MIT, for instance) but other than that this is
probably a pretty complete collection of files.

Note that this does not include notably specific files such as
project.json, .gitignore, and others. We may want to add those,
but we may not, and just create forks of this project, that are
effectively variants. (E.g. for a node project, it might be
zambezi/project-node, or some such.)