Add initialization script for easier project generation
This script allows people to simply clone this repo and run the script. I debated doing the `curl <script-url> | bash` thing that's popular these days, but decided against it since it's just not worth the hassle of dealing with things like partial downloads and the likes. By using git to clone the repo and only if that succeeds run the script we ensure we don't run into situations where the script will merrily run (and fail miserably) even if not all of it is downloaded.